Separate rustfmt-config
