+++ rustfmt-config/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "rustfmt_config"
+version = "0.1.0"
+authors = ["topecongiro <seuchida@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ansi_term = "0.12"
+dirs = "2.0"
+itertools = "0.8"
+regex = "1.3"
+rustfmt-config_proc_macro = { version = "0.2", path = "config_proc_macro" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
+toml = "0.5"
+
+[dependencies.syntax]
+package = "rustc-ap-syntax"
+version = "610.0.0"
+
+[dependencies.syntax_pos]
+package = "rustc-ap-syntax_pos"
+version = "610.0.0"
+
