+++ rustfmt-config/src/config_type.rs
@@ -1,5 +1,5 @@
-use crate::config::file_lines::FileLines;
-use crate::config::options::{IgnoreList, WidthHeuristics};
+use crate::file_lines::FileLines;
+use crate::options::{IgnoreList, WidthHeuristics};
 
 /// Trait for types that can be used in `Config`.
 pub(crate) trait ConfigType: Sized {
@@ -52,7 +52,6 @@ impl ConfigType for IgnoreList {
 
 macro_rules! create_config {
     ($($i:ident: $Ty:ty, $def:expr, $is_stable:literal, $dstring:literal;)+) => (
-        #[cfg(test)]
         use std::collections::HashSet;
         use std::io::Write;
 
@@ -155,7 +154,7 @@ macro_rules! create_config {
                         self.$i.1 = true;
                         self.$i.2 = val;
                     } else {
-                        if crate::is_nightly_channel!() {
+                        if is_nightly_channel!() {
                             self.$i.1 = true;
                             self.$i.2 = val;
                         } else {
@@ -172,15 +171,14 @@ macro_rules! create_config {
             }
 
             /// Returns a hash set initialized with every user-facing config option name.
-            #[cfg(test)]
-            pub(crate) fn hash_set() -> HashSet<&'static str> {
+            pub fn hash_set() -> HashSet<&'static str> {
                 [$(
                     stringify!($i),
                 )+]
                 .iter().copied().collect()
             }
 
-            pub(crate) fn is_valid_name(name: &str) -> bool {
+            pub fn is_valid_name(name: &str) -> bool {
                 match name {
                     $(
                         stringify!($i) => true,
@@ -344,4 +342,10 @@ macro_rules! create_config {
             }
         }
     )
+}
+
+macro_rules! is_nightly_channel {
+    () => {
+        option_env!("CFG_RELEASE_CHANNEL").map_or(true, |c| c == "nightly" || c == "dev")
+    };
 }
