+++ rustfmt-config/src/file_lines.rs
@@ -36,6 +36,15 @@ impl From<source_map::FileName> for FileName {
     }
 }
 
+impl From<&FileName> for syntax_pos::FileName {
+    fn from(filename: &FileName) -> syntax_pos::FileName {
+        match filename {
+            FileName::Real(path) => syntax_pos::FileName::Real(path.to_owned()),
+            FileName::Stdin => syntax_pos::FileName::Custom("stdin".to_owned()),
+        }
+    }
+}
+
 impl fmt::Display for FileName {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
@@ -197,12 +206,12 @@ fn normalize_ranges(ranges: &mut HashMap<FileName, Vec<Range>>) {
 
 impl FileLines {
     /// Creates a `FileLines` that contains all lines in all files.
-    pub(crate) fn all() -> FileLines {
+    pub fn all() -> FileLines {
         FileLines(None)
     }
 
     /// Returns `true` if this `FileLines` contains all lines in all files.
-    pub(crate) fn is_all(&self) -> bool {
+    pub fn is_all(&self) -> bool {
         self.0.is_none()
     }
 
@@ -250,22 +259,22 @@ impl FileLines {
 
     /// Returns `true` if `range` is fully contained in `self`.
     #[allow(dead_code)]
-    pub(crate) fn contains(&self, range: &LineRange) -> bool {
+    pub fn contains(&self, range: &LineRange) -> bool {
         self.file_range_matches(&range.file_name(), |r| r.contains(Range::from(range)))
     }
 
     /// Returns `true` if any lines in `range` are in `self`.
-    pub(crate) fn intersects(&self, range: &LineRange) -> bool {
+    pub fn intersects(&self, range: &LineRange) -> bool {
         self.file_range_matches(&range.file_name(), |r| r.intersects(Range::from(range)))
     }
 
     /// Returns `true` if `line` from `file_name` is in `self`.
-    pub(crate) fn contains_line(&self, file_name: &FileName, line: usize) -> bool {
+    pub fn contains_line(&self, file_name: &FileName, line: usize) -> bool {
         self.file_range_matches(file_name, |r| r.lo <= line && r.hi >= line)
     }
 
     /// Returns `true` if all the lines between `lo` and `hi` from `file_name` are in `self`.
-    pub(crate) fn contains_range(&self, file_name: &FileName, lo: usize, hi: usize) -> bool {
+    pub fn contains_range(&self, file_name: &FileName, lo: usize, hi: usize) -> bool {
         self.file_range_matches(file_name, |r| r.contains(Range::new(lo, hi)))
     }
 }
