+++ rustfmt-config/src/license.rs
