+++ rustfmt-core/Cargo.toml
@@ -18,7 +18,7 @@ itertools = "0.8"
 lazy_static = "1.0.0"
 log = "0.4"
 regex = "1.0"
-rustfmt-config_proc_macro = { version = "0.2", path = "../config_proc_macro" }
+rustfmt_config = { path = "../rustfmt-config" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 term = "0.6"
